requirements.txt
imap-tools
mysql-connector-python
pdfplumber
